requirements.txt
astroid==2.4.2
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
dateparser==0.7.6
Flask==1.1.2
Flask-SocketIO==4.3.1
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
regex==2020.10.11
requests==2.24.0
service-identity==18.1.0
six==1.15.0
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.2
